three fixes to the directory code:
 - VFlushVnode is called when rvm transactions abort.
   Aborting with a newly created directory had a bug
   leading to an assertion because of a refcount
 - VN_CopyOnWrite had a bug in the RVM refcount for
   directories.  This has unfortunately led to leaked
   RVM that we can only recover by reinitializing
 - Repair code in the server: a few directory handles
   were gotten and not put away. Fixed.

Let's try this!

- Peter -

+++ lib-src/kernel-includes/coda.h
@@ -241,7 +241,7 @@ struct coda_cred {
 enum coda_vtype	{ C_VNON, C_VREG, C_VDIR, C_VBLK, C_VCHR, C_VLNK, C_VSOCK, C_VFIFO, C_VBAD };
 
 struct coda_vattr {
-	int     	va_type;	/* vnode type (for create) */
+	long     	va_type;	/* vnode type (for create) */
 	u_short		va_mode;	/* files access mode and type */
 	short		va_nlink;	/* number of references to file */
 	vuid_t		va_uid;		/* owner user id */
